refactor: move memory delivery outbox out of economy module
- new src/engine/memory_outbox.py owns the five delivery/reversal state
  functions; MAX_EXTERNAL_EFFECT_DELIVERIES moves with it (breaks the
  economy<->outbox import cycle); data stays in economy.external_effects_outbox
  because it rides economy persistence and rollback
- effect_groups triage: kept in economy.py -- it is the multi-proposal
  all-or-nothing effect barrier used by round_processor / game_lifecycle /
  swipe_generator, not memory-only state (docstring already documents this)
- economy_effects triage: kept -- pure stateless narration guard / notice /
  reward-eligibility helpers consumed by the round pipeline; no own state
  machine or persistence fields; stale purchase-order notice copy updated
